docs: note citizen helpers
